Update deleting-a-lead-or-contact.md
According to an Adobe PM, if a record is deleted in Marketo, and the associated record in MSD is updated, then the record will not be recreated in Marketo. The reason is because the connector in Dynamics is looking for UPDATE to an existing record which doesn’t exist any longer, instead of a 'create' action. See Support Ticket 02158061 for documentation of this expected behavior -- which is the opposite of what was in the documentation.